==> 1_B_quadrant.R <==
library(Matrix)
library(parallel)


# Matrices necessary

sup <- read.csv("fabio-gloria_sup.csv")
use <- read.csv("fabio-gloria_use.csv")
conc <- read.csv("fabio-gloria_conc.csv")
conc$FABIO_code <- 1:nrow(conc)
# Move NAs to extra column to drop, allocate at some point
conc$GLORIA_code[is.na(conc$GLORIA_code)] <- 165

as_matrix <- function(x) {
  y <- as.matrix(x[5:ncol(x)])
  y_rowSums <- rowSums(y, na.rm = TRUE)
  if(!all(x[["Total"]] == y_rowSums)) stop()
  y[is.na(y)] <- 0
  dimnames(y) <- NULL
  Matrix(y)
}
Sup <- as_matrix(sup)
Use <- as_matrix(use)
Cou_NA <- sparseMatrix(i = conc$FABIO_code, j = conc$GLORIA_code) * 1
Cou <- Cou_NA[, 1:164] # Remove 165th column of countries missing in GLORIA

# Function to create the hybrid part for a certain year
hybridise <- function(year, Sup, Use, Cou, Y_all) {
  
  require(Matrix) # Necessary for forked processes
  # Read GLORIA Z and FABIO Y
  Y <- Y_all[[as.character(year)]]
  load(paste0(paste0("gloria/EEMRIO_rdata/",year,"_Z.RData")))
  
  # Calculate Tech matrices for the 164 gloria countries
  Tec <- vector("list", 164)
  for(i in 1:164) {
    tmp <- Matrix(0, nrow = 120, ncol = 120)#120 GLORIA products
    for(j in 1:164)
      tmp <- tmp + Z[(1 + 120 * (j - 1)):(120 * j), 
                     (1 + 120 * (i - 1)):(120 * i)]
    Tec[[i]] <- tmp 
  }
  
  # Get the columns of Y containing the other use category to allocate
  Oth <- Y[, grep("other$", colnames(Y))]
  rm(Y, Z); gc()
  
  # Match FABIO countries with GLORIA countries and restructure the Other use matrix
  Oth <- Oth %*% Cou
  nprod <- nrow(Oth) / 192#192 FABIO countries
  
  # Create matrix for sector matching
  T <- vector("list", 164)
  for(i in 1:164) {
    T[[i]] <- Sup[1:nprod,] %*% Tec[[i]] * Use[1:nprod,]
    T[[i]] <- T[[i]] / rowSums(T[[i]])
    T[[i]][is.na(T[[i]])] <- 0
  }
  
  # Compute the hybrid part from Oth and T
  B <- Matrix(0, nrow = 192 * nprod, ncol = 164 * 120, sparse = TRUE)
  for(i in 1:164) {
    B[, (1 + 120 * (i - 1)):(120 * i)] <-
      do.call(rbind, replicate(192, T[[i]], simplify = FALSE)) * Oth[, i]
  }
  rm(T); gc()
  B
  
}




# Execute -----------------------------------------------------------------

# Select fabio version or run loop
#versions <- c("", "losses/", "wood/")
#version <- versions[1]

#for(version in versions){
Y_all <- readRDS(paste0("fabio_v1.2/Y.rds"))

# Setup to process in parallel
n_cores <- parallel::detectCores() - 2
cl <- parallel::makeCluster(n_cores)

# Years to calculate hybridised FABIO for
#years <- 1986:2013
years <- 1990:2020

output <- parallel::parLapply(cl, years, hybridise, Sup, Use, Cou, Y_all)

for(i in seq_along(output)) {
  saveRDS(output[[i]], 
          paste0("fabio_v1.2/hybrid/", years[[i]], "_B.rds"))
}

parallel::stopCluster(cl)

rm(cl, Y_all, n_cores, years, output); gc()

# # Alternatively run a loop
# for(year in years){
#   saveRDS(hybridise(year, Sup, Use, Cou, Y_all), 
#           paste0("/mnt/nfs_fineprint/tmp/fabio/v2/hybrid/", version, year, "_B.rds"))
# }
#}
